# apiserver/openapi.yaml
openapi: 3.0.1
info:
  title: APIRecintos
  description: API para prestação de informações sobre eventos de controle aduaneiro
    a cargo dos Redex, Recintos, Administradores e Operadores Aeroportuários, Portuários
    e demais intervenientes em carga sobre controle aduaneiro.
  version: "1.0b"
servers:
- url: /
paths:
  /auth:
    post:
      summary: retorna token JWT
      operationId: authentication.generate_token
      requestBody:
        description: Terminal e senha
        content:
          application/json:
            schema:
              x-body-name: recinto
              $ref: '#/components/schemas/AuthenticationJson'
      responses:
        '200':
          description: JWT token
          content:
            'text/plain':
              schema:
                type: string
  /secret:
    get:
      summary: Return secret string
      operationId: authentication.get_secret
      responses:
        '200':
          description: secret response
          content:
            'text/plain':
              schema:
                type: string
      security:
      - jwt: ['secret']
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: authentication.decode_token
  schemas:
    AuthenticationJson:
      type: object
      properties:
        recinto:
          type: string
        senha:
          type: string
...